refactor:#166 layout 경로 수정
